requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.10
Django==5.2
django-crispy-forms==2.4
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
